+++ include/Planet.h
@@ -7,7 +7,7 @@ class Planet : public SpaceObject {
 public:
     static const float DEFAULT_DENSITY;
     Planet(const sf::Texture &texture, const sf::Vector2f& pos, const sf::Vector2f& vel,
-           float rot, float radius);
+           float rot, int radius);
 };
 
 #endif //SPACE_SHOOTER_PLANET_H
