+++ include/SpaceObject.h
@@ -8,7 +8,7 @@ public:
     static const float G_;
     SpaceObject(const sf::Texture &texture, const sf::Vector2f& initPos,
                 const sf::Vector2f& initVel, float initRot,
-                float radius, float density);
+                int radius, float density);
 
     void update();
     float getMass() const;
@@ -16,7 +16,7 @@ public:
 private:
     float density_;
     float mass_;
-    float calculateMass(float radius);
+    float calculateMass(int radius);
 };
 
 #endif //SPACE_SHOOTER_SPACEOBJECT_H
